requirements.txt
absl-py==0.9.0
astunparse==1.6.3
Brotli==1.0.7
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dash==1.14.0
dash-bootstrap-components==0.10.3
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
elasticsearch==7.8.0
fastavro==0.24.0
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gast==0.3.3
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
multitasking==0.0.9
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.0.5
plotly==4.9.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.8.0
termcolor==1.1.0
textblob==0.15.3
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yfinance==0.1.54
zipp==3.1.0
